+++ packages/ariakit-test/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @ariakit/test
 
+## 0.4.12
+
+- Updated packages to target ES2018 (previously ES2017).
+- Updated dependencies: `@ariakit/core@0.4.17`
+
 ## 0.4.11
 
 - Updated dependencies: `@ariakit/core@0.4.16`